tekstil-lead-finder/requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
httpx>=0.25.0
fake-useragent>=1.4.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0

# Email Tools
email-validator>=2.1.0
dnspython>=2.4.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Dashboard
streamlit>=1.29.0
plotly>=5.18.0

# Scheduling & Async
schedule>=1.2.0
asyncio>=3.4.3
aiohttp>=3.9.0

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Rate Limiting
ratelimit>=2.2.1

# CLI
rich>=13.7.0
typer>=0.9.0
